Feat/manual UI

+++ app/src/main/java/com/zenefit/zenefit_android/data/remote/reqeust/RequestSignUpData.kt
@@ -5,7 +5,7 @@ data class RequestSignUpData(
     val age : Int,
     val areaCode : String,
     val cityCode : String,
-    val lastYearIncome : Int,
+    val lastYearIncome : Double,
     val educationType : String,
     val jobs : List<String>,
     val marketingStatus : Boolean
